+++ gsplus-main/gsplus/src/compile_time.c
@@ -1,2 +1,0 @@
-char g_compile_time[] = "Compiled: " __DATE__ " " __TIME__ ;
-