+++ gsplus-main/upstream/kegs/src/Kegs-Bridging-Header.h
@@ -1,5 +1,0 @@
-// $KmKId: Kegs-Bridging-Header.h,v 1.1 2019-10-14 22:33:09+00 kentd Exp $
-//  Use this file to import your target's public headers that you would like to expose to Swift.
-//
-
-#import "defc.h"